Move schema pkg -> json/schema
