+++ internal/pkg/json/schema/generate.go
