+++ internal/pkg/json/schema/generate_test.go
