+++ internal/pkg/json/schema/schema.go
