+++ internal/pkg/json/schema/schema_test.go
