+++ internal/pkg/local/create.go
@@ -5,8 +5,9 @@ import (
 
 	"github.com/iancoleman/orderedmap"
 
+	"github.com/keboola/keboola-as-code/internal/pkg/json/schema"
+	"github.com/keboola/keboola-as-code/internal/pkg/json/schema"
 	"github.com/keboola/keboola-as-code/internal/pkg/model"
-	"github.com/keboola/keboola-as-code/internal/pkg/schema"
 )
 
 func (m *Manager) CreateObject(key model.Key, name string) (model.Object, error) {
